Rename states, windows, dialogs and files
