+++ lab_4/src/bot/dialogs/main_menu/__init__.py
@@ -1,18 +1,18 @@
 from aiogram_dialog import Dialog
 
-from .add_task import window as add_task_window
-from .main_menu import window as main_menu_window
-from .select_task import window as select_task_window
-from .delete_or_edit_task import window as delete_or_edit_task_window
-from .delete_task import window as delete_task_window
-from .edit_task import window as edit_task_window
+from .choosing_action import window as choosing_action_window
+from .adding_task import window as adding_task_window
+from .choosing_task import window as choosing_task_window
+from .task_choosen import window as task_choosen_window
+from .deleting_task import window as deleting_task_window
+from .editing_task import window as editing_task_window
 
 
 dialog = Dialog(
-    add_task_window,
-    main_menu_window,
-    select_task_window,
-    delete_or_edit_task_window,
-    delete_task_window,
-    edit_task_window,
+    choosing_action_window,
+    adding_task_window,
+    choosing_task_window,
+    task_choosen_window,
+    deleting_task_window,
+    editing_task_window,
 )
